refactor: car 관련 도메인들 car 패키지로 분리
