+++ src/test/kotlin/racingcar/domain/RacingResultTest.kt
@@ -2,6 +2,10 @@ package racingcar.domain
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.car.Car
+import racingcar.domain.car.Cars
+import racingcar.domain.car.Name
+import racingcar.domain.car.Position
 
 class RacingResultTest : StringSpec({
     "우승자를 찾는 기능" {
