+++ src/test/kotlin/racingcar/domain/car/CarTest.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.domain.car
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.data.row
